basic blocks: use char index instead of int

+++ compiler/src/org.graalvm.compiler.core.common/src/org/graalvm/compiler/core/common/cfg/Loop.java
@@ -63,7 +63,7 @@ public abstract class Loop<T extends AbstractBlockBase<T>> {
         this.naturalExits = new ArrayList<>();
     }
 
-    public abstract long numBackedges();
+    public abstract int numBackedges();
 
     @Override
     public String toString() {
